requirements.txt
streamlit>=1.28.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Hugging Face Embeddings (Local, Free)
sentence-transformers>=2.2.0
torch>=2.0.0
transformers>=4.30.0

# Google Gemini LLM (Generous Free Tier)
google-generativeai>=0.3.0

# Text processing
langchain-text-splitters>=0.0.1

# HTTP requests
requests>=2.31.0

# Logging and utilities
colorama>=0.4.6
tqdm>=4.65.0

# Data processing
pathlib
json5>=0.9.0

# Optional: for better performance
accelerate>=0.20.0
optimum>=1.12.0 